Acceleration by Perlin noise
The time series has its own "acceleration", the difference between
successive values of noise consititutes a rate of change. Given the time
series differences represent acceleration already, this version uses
that acceleration directly as the Mover's acceleration. There is no call
to map.

I expected this to be faster to calculate, and so smoother, but I'm not
convinced. I think the leaps might be larger, giving a less-smooth
appearance. I could play with different index steps.